Fix Windows conversion warnings in misc.cpp, and a dynamic memory allocation disguised as a stack array.

+++ Shared/misc.cpp
@@ -25,11 +25,11 @@ namespace shared {
 size_t formatDate(char* buf, size_t const max, int64_t const unixtime) {
   DivUMod const div_day = divUMod(unixtime, 24 * 60 * 60);
   DivUMod const div_era = divUMod(div_day.quot - 11017, 146097);
-  unsigned const doe = div_era.rem;
+  unsigned const doe = static_cast<unsigned>(div_era.rem);
   unsigned const yoe = (doe - doe / 1460 + doe / 36524 - (doe == 146096)) / 365;
   unsigned const doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
   unsigned const moy = (5 * doy + 2) / 153;
-  int const y = 2000 + div_era.quot * 400 + yoe + (9 < moy);
+  int const y = static_cast<int>(2000 + div_era.quot * 400 + yoe + (9 < moy));
   unsigned const m = moy + (9 < moy ? -9 : 3);
   unsigned const d = doy - (153 * moy + 2) / 5 + 1;
   int const len = snprintf(buf, max, "%04d-%02u-%02u", y, m, d);
@@ -48,14 +48,14 @@ size_t formatDateTime(char* buf,
   DivUMod const div_hip = divUMod(timestamp, pow10[dimension]);
   DivUMod const div_day = divUMod(div_hip.quot, 24 * 60 * 60);
   DivUMod const div_era = divUMod(div_day.quot - 11017, 146097);
-  unsigned const doe = div_era.rem;
+  unsigned const doe = static_cast<unsigned>(div_era.rem);
   unsigned const yoe = (doe - doe / 1460 + doe / 36524 - (doe == 146096)) / 365;
   unsigned const doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
   unsigned const moy = (5 * doy + 2) / 153;
-  int const y = 2000 + div_era.quot * 400 + yoe + (9 < moy);
+  int const y = static_cast<int>(2000 + div_era.quot * 400 + yoe + (9 < moy));
   unsigned const m = moy + (9 < moy ? -9 : 3);
   unsigned const d = doy - (153 * moy + 2) / 5 + 1;
-  unsigned const minutes = div_day.rem / 60;
+  unsigned const minutes = static_cast<unsigned>(div_day.rem) / 60;
   unsigned const ss = div_day.rem % 60;
   unsigned const hh = minutes / 60;
   unsigned const mm = minutes % 60;
@@ -76,7 +76,7 @@ size_t formatDateTime(char* buf,
 }
 
 size_t formatHMS(char* buf, size_t const max, int64_t const unixtime) {
-  unsigned const seconds = unsignedMod(unixtime, 24 * 60 * 60);
+  unsigned const seconds = static_cast<unsigned>(unsignedMod(unixtime, 24 * 60 * 60));
   unsigned const minutes = seconds / 60;
   unsigned const ss = seconds % 60;
   unsigned const hh = minutes / 60;
